requirements.txt
easyocr
requests
Pillow
colorama
pyautogui